windows build complains about conversion

+++ PongMode.cpp
@@ -328,7 +328,7 @@ void PongMode::draw(glm::uvec2 const &drawable_size) {
 				cf = 0.0f;
 			}
 			if (ci > int32_t(trail_colors.size())-1) {
-				ci = trail_colors.size()-1;
+				ci = int32_t(trail_colors.size())-1;
 				cf = 1.0f;
 			}
 			//do the interpolation (casting to floating point vectors because glm::mix doesn't have an overload for u8 vectors):
